Draft for reloadable configuration (Addressing oatpp/oatpp#483).
`ssl_ctx_st` are now bound to their respective `Config` via shared pointers.
When creating an `Connection`, the config is passed and stored in the `Connection`, also in a shared pointer.
Thus, we have full reference counting on the `ssl_ctx_st`.
The context is automatically freed when no `Connection` exists with the config containing the context and the `ConnectionProvider` has loaded a new config.

Discussion: Since the context is now part of the config, shouldn't the code in `instantiateTLSServer` moved to `Config`? This way, SSLv23_server_method would be made easily configurable, too.